+++ src/main/java/modelo/InterfazPosicionObjeto.java
@@ -1,5 +1,5 @@
 package modelo;
 
 public interface InterfazPosicionObjeto {
-    void ubicarPosicionTablero(int x, int y);
+    void ubicarPosicionTablero(double x, double y);
 }
